feat: work on update user
