+++ src/app/api/users/[id]/route.ts
@@ -1,7 +1,11 @@
 import { revalidateTag } from "next/cache";
 import { cookies } from "next/headers";
 
-export async function POST(req: Request) {
+export async function PUT(
+  req: Request,
+  { params }: { params: Promise<{ id: number }> },
+) {
+  const id = (await params).id;
   const data = await req.json();
   const cookieStore = cookies();
   const token = cookieStore.get("jwt")?.value;
@@ -10,23 +14,24 @@ export async function POST(req: Request) {
     return Response.json({ error: "No token, Unauthorized" }, { status: 401 });
   }
 
+  console.log(data);
   const res = await fetch(
-    `${process.env.NEXT_PUBLIC_STRAPI_URL}api/services`,
+    `${process.env.NEXT_PUBLIC_STRAPI_URL}api/users/${id}`,
 
     {
       headers: {
         "Content-Type": "application/json",
         Authorization: `Bearer ${token}`,
       },
-      method: "POST",
+      method: "PUT",
       body: JSON.stringify({ data }),
-      next: {
-        tags: ["services"],
-      },
       cache: "no-cache",
+      next: {
+        tags: ["me"],
+      },
     },
   );
-  revalidateTag("services");
+  revalidateTag("me");
 
   const value = await res.json();
   if (!res.ok) {
