+++ src/app/api/works/[id]/route.ts
@@ -25,7 +25,6 @@ export async function PUT(
       method: "PUT",
       body: JSON.stringify({ data }),
       next: {
-        revalidate: 15,
         //tags: ["works", `work-${id}`],
       },
     },
